move table to separate file
