+++ corehq/apps/integration/kyc/models.py
@@ -2,9 +2,7 @@ from django.db import models
 from django.utils.translation import gettext as _
 
 import jsonfield
-from django_tables2 import columns
 
-from corehq.apps.hqwebapp.tables.htmx import BaseHtmxTable
 from corehq.motech.models import ConnectionSettings
 
 
@@ -35,47 +33,3 @@ class KycConfig(models.Model):
         constraints = [
             models.UniqueConstraint(fields=['domain', 'provider'], name='unique_domain_provider'),
         ]
-
-
-class KycVerifyTable(BaseHtmxTable):
-    class Meta(BaseHtmxTable.Meta):
-        pass
-
-    verify_select = columns.CheckBoxColumn(
-        accessor='id',
-        attrs={
-            'th__input': {'name': 'select_all'},
-            'td__input': {'name': 'selection'}
-        }
-    )
-    first_name = columns.Column(
-        verbose_name=_("First Name"),
-    )
-    last_name = columns.Column(
-        verbose_name=_("Last Name"),
-    )
-    phone_number = columns.Column(
-        verbose_name=_("Phone Number"),
-    )
-    email = columns.Column(
-        verbose_name=_("Email Address"),
-    )
-    national_id_number = columns.Column(
-        verbose_name=_("National ID Number"),
-    )
-    street_address = columns.Column(
-        verbose_name=_("Street Address"),
-    )
-    city = columns.Column(
-        verbose_name=_("City"),
-    )
-    post_code = columns.Column(
-        verbose_name=_("Post Code"),
-    )
-    country = columns.Column(
-        verbose_name=_("Country"),
-    )
-    verify_btn = columns.TemplateColumn(
-        template_name='kyc/partials/kyc_verify_button.html',
-        verbose_name=_("Verify"),
-    )
